requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
supabase==2.4.6
python-dotenv==1.0.1
schedule==1.2.1
lxml==5.2.1
fake-useragent==1.5.1
tenacity==8.2.3
playwright==1.44.0
